pythonServer/requirements.txt
Flask==3.0.3
Flask_Cors==4.0.1
numpy==1.26.4
pandas==2.2.2
psycopg2_binary==2.9.9
scikit_learn==1.4.2
sentence_transformers==2.7.0
spacy==3.7.4
